Remove Maven build config, fix Gradle build config
This removes the Maven pom.xml from when I first created the project with
Maven. There's no need to keep it around, and it can confuse IDE's, so
it's best to remove it.

I also added an entry in .gitignore for the /logs/ directory, which is
created when unit tests are run.

Finally, I changed a few things in the build.gradle file:

 1. `testCompileOnly 'org.apiguardian:apiguardian-api:1.0.0'` - removes
    the warning at compile time regarding an apiguardian class that can't
    be found at test time. apiguardian is used by Junit 5.

 2. `testCompile sourceSets.mlservice.output` - was testCompileOnly, but
    tests fail without it.

 3. I formatted all the dependencies to use the same style.

 4. `compile sourceSets.antlr4.output` - this didn't seem to affect the
    Gradle build, but it did confuse IntelliJ without it.

 5. `antlr4Compile 'org.antlr:antlr4:4.7.1'` - was just `antlr4`, which
    worked fine, but IntelliJ couldn't find the antlr4 sources in the
    generated source set. This lead to invalid errors where the
    generated sources were used.

 6. `classpath = configurations.antlr4Compile` - because of #5, the
    classpath entry in `generateGrammarSource` needed to be updated.