Fail build on missing or empty config
To enforce the usage of AccessTransformers when requested, the build
will now fail if the requested AccessTransformers config file is not
defined, cannot be found, or is empty.

This validation will happen after project evaluation rather than at
artifact transformation, because exceptions thrown by artifact
transforms are consumed and don't fail the build.

Also, added `AccessTransformerContainer#setLogLevel` to allow for
viewing the output stream of AccessTransformers using a different log
level.