Use dependency substitution rules instead of constraints
Dependency substitutions are only considered and applied for incoming
configurations. Consumable configurations are not affected and thus the
attributes used to enforce AccessTransformers will not pollute the
Gradle Module Metadata.

This is important because it is theoretically possible for stray
AccessTransformer attributes to either:
1. Cause a dependency lock-up for dependencies that have weird variants
or complex resolutions.
2. Cause an incoming dependency with those attributes to have unwanted
AccessTransformers applied to them.