Fix ATs causing IDEs to lose source jars
This was done by making the access transformer only target dependencies of type Usage: "library". Sources are of type Usage: "documentation".

This problem is two-fold:
1. Gradle would access transform the sources, but wouldn't select it as an auxiliary artifact for the IDE.
2. IntelliJ would look for the auxiliary artifact but can't find it due to the variant change from Gradle transforming the artifact.

This is a problem that would be fixed by gradle/gradle#35131. When it is, we can theoretically add support in AccessTransformers itself to transform `.java` files as well as `.class` files.