Attempt to inform consumer if transformation failed
Exceptions that are thrown in artifact transforms are swallowed by Gradle unless it's necessary for a task execution. This means that artifact transforms that are run as part of IDE sync may have its exceptions silently ignored, while the dependency fails to resolve.

The workaround for now is simply to try and log the thrown exception before throwing it, by wrapping the transform process in a try/catch.